backend/pyproject.toml
[project]
name = "spendhelm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "fastapi-cors>=0.0.6",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "passlib[argon2]>=1.7.4",
    "psycopg[binary]>=3.3.2",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "slowapi>=0.1.9",
    "fastapi-pagination>=0.15.4",
    "asyncpg>=0.31.0",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "fastapi.Depends",
    "fastapi_pagination.limit_offset.LimitOffsetParams"
]